Rename module to rir_builder
